+++ doc/code/rust/src/disassembler.rs
@@ -42,3 +42,17 @@ pub fn dwarf_function() {
     }
     // lief-doc: dwarf-func-end
 }
+
+pub fn x86_lock(some_inst: &lief::assembly::x86::Instruction) {
+    // lief-doc: x86-lock-start
+    let inst: &lief::assembly::x86::Instruction = some_inst;
+
+    if inst.has_lock_prefix() || inst.is_atomic() {
+        println!("{} is atomic", inst);
+    } else if inst.is_lockable()
+        && let Some(locked) = inst.lock()
+    {
+        println!("atomic version: {}", locked);
+    }
+    // lief-doc: x86-lock-end
+}